+++ include/bccboardrenderer.h
@@ -17,14 +17,16 @@ struct AtlasImageOps
 {
     int opType;
     int id    ;
-    int posX  ;
-    int posY  ;
+    qreal posX  ;
+    qreal posY  ;
     int tX    ;
     int tY    ;
     int tW    ;
     int tH    ;
 };
 
+
+//all should be int ...
 struct BrushCall{
     qreal cellX;
     qreal cellY;
